refactor: gate agents page redesign behind a feature switch
- Add the AgentsPageRedesign feature switch (off by default).
- Keep the original two-section layout as AgentsPageLegacy (default) and
  move the tabbed redesign into AgentsPageTabs; the AgentsPage entry now
  renders one or the other based on the switch.
- Restore the legacy integration tests and add a separate suite that
  exercises the redesign with the switch enabled.
